+++ packages/sdk-core/CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 - `SuperToken` class has `send` function
 - `batchCall` supports new `send` batch operation
+- Added `downgradeTo` function on `SuperToken` class
+- Added toga events to be part of query
 
 ## [0.5.9] - 2022-12-05